Attempt to improve email legibility on phones
Some of the provider names are quite long and wrap when phones are held
in portrait orientation (as is most common) — especially when someone is
using a larger font size. Due to the wrapping the providers start to run
together, i.e. it’s hard to make them out. I hope this helps.

I hope this fixes #28 but I haven’t tested it. Too tired.
Here’s hoping!